Refactor for type safety
Unfortunately, in order to get the type assertions working I had to move
the formatting of the messages into the loop and use scope variables.
The type defs don't allow for a function argument to be passed to the
message being formatted.

This required me to refactor how the tests worked as well. They now
expect an explicit string instead of a value passed by reference using
exposed messages. This is slightly more brittle — if Stylelint changes
the way they format messages, this will break.